pyproject.toml
[project]
name = "frappe_deep_agents"
version = "0.0.1"
description = "LangChain Deep Agents implementation for Frappe with Kubernetes sandbox support"
authors = [
    { name = "Frappe Accelerated", email = "admin@frappe-accelerated.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "langchain>=0.3",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-ollama>=0.2",
    "kubernetes>=29.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0"
